Remove files generated by cmake.
Generally, the git repository should only contain original source files, not anything that's
been generated by the compiler or by cmake or any other external tool. That's because if
you add these generated files, then a small change in one source file might entail a large
change in generated files and you're going to bloat your repository with stuff that doesn't
actually have to be stored in the repository.